Small Code Refactor (Check Desc)
- Merged all deserialization into one function inside of the UnrealArchive class. With the use of generics, its much more clean now.
- Merged UDynamicArrayManager into UPropertymanager. A seperate folder that contains UDynamicArrayManager is unecessary.
- Refactored UPropertyManager so its code is more concise and readable.
- Removed some folders that were uneeded

There is still a LOT more to do in terms of refactoring. This was just a needed cleanup before moving onto deserializing static uprop arrays. Once that is finished I will move onto serializing the newly formated data back in: this time around it will be a lot more optimized.